Unset max width for notice images.
Some themes have max-width: 100% set to prevent images from breaking
out of containers. The notice images tend to be close to square, so it
shouldn't be an issue to unset the max-width. Max-width: 100% was a
particular problem for images wrapped in links. This is in response to
an extremely edge case usage of MetadataBrowse being applied to a
property using a notice.